bacon: strip absolute PRODUCT_OUT path from md5sum file
When lineage-$(CM_VERSION).zip.md5sum is created, md5sum is passed the
absolute path, which ends up in the md5sum file. This absolute path is
obviously incorrect on the device when flashing the ROM.

This patch runs md5sum through sed to strip out the "$(PRODUCT_OUT)/"
text from the resulting md5sum file, allowing it to be copied to the
device and verified by TWRP etc. when flashing.

Change-Id: Icd54889a8168ed8fc3f7d7548267100b96549eee
Signed-off-by: Allen Wild <allenwild93@gmail.com>
Signed-off-by: klozz <xxx.reptar.rawrr.xxx@gmail.com>